requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
asterism==0.7.2
    # via -r requirements.in
attrs==21.2.0
    # via jsonschema
bagit==1.8.1
    # via
    #   -r requirements.in
    #   asterism
    #   bagit-profile
bagit-profile==1.3.1
    # via -r requirements.in
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
clinner==1.12.3
    # via health-check
colorlog==3.2.0
    # via clinner
django==3.2.7
    # via
    #   -r requirements.in
    #   asterism
    #   djangorestframework
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   asterism
gitdb==4.0.7
    # via gitpython
gitpython==3.1.18
    # via health-check
health-check==3.4.1
    # via -r requirements.in
idna==2.10
    # via requests
importlib-metadata==4.6.1
    # via jsonschema
jsonschema==3.2.0
    # via rac-schemas
odin==1.7.0
    # via asterism
psycopg2-binary==2.9.1
    # via asterism
pyrsistent==0.18.0
    # via jsonschema
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via health-check
rac_schemas==0.29
    # via -r requirements.in
requests==2.25.1
    # via bagit-profile
six==1.16.0
    # via
    #   jsonschema
    #   odin
smmap==4.0.0
    # via gitdb
sqlparse==0.4.1
    # via django
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   gitpython
    #   importlib-metadata
urllib3==1.26.6
    # via requests
zipp==3.5.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
